move header component to organisms
